Replaced handler loading commands such that what was
php_value uri_handler /tmp/foo.php

is now

phpUriHandler	/tmp/foo.php

This fixes some bugs that caused handlers to be skipped or mysteriously 'unloaded',
reduces the number of calls to zend_alter_ini_entries that are necessary, as well
as allows for easier implementation of stacked handlers (which comes next).

Added 2 new hooks, phpPostReadHandler and phpHeaderHandler (going to the obvious places).


git-svn-id: http://svn.php.net/repository/php/php-src/branches/apache_hooks@93815 c90b9560-bf6c-de11-be94-00142212c4b1